Henry changes single opponent and deal with draws.
The opponent attribute of State is now a single Bike object
rather than a list of Bikes. Fair few simplifying changes to deal
with this. Reward now deals with draws correctly.

Jimmy made a few changes to henry's sent file to fix the arena size
and put the barriers back. Also added back in screen injection to
render_game.