Ir2 changes (#2)
* readme up to date with ir1

* add henrys mstc stuff with some small changes

* got simulation from base to pass, may not 100% be correct but test pass

* addded mcts_main

* added back last_action into Node

* looks like everything working, snake playing decently against itself

* working with the dont die rollout policy

* now compute visual constants from injected screen if given

* minor updates pre henry changes

* henry changes for mcts

* henry changes for mcts

* re-add screen injection

* Henry changes single opponent and deal with draws.

The opponent attribute of State is now a single Bike object
rather than a list of Bikes. Fair few simplifying changes to deal
with this. Reward now deals with draws correctly.

Jimmy made a few changes to henry's sent file to fix the arena size
and put the barriers back. Also added back in screen injection to
render_game.

* Minor tweaks to main.py given new format

Co-authored-by: neuromantic99 <jrowlandm@gmail.com>